fix: removed user-input controller abstract class and derived ones
