--- packages/starknet-snap/src/user-inputs/index.tsx
@@ -1,1 +1,0 @@
-export * from './fee-token-selector';
